fuzz: add basic evaluation script
When generating inputs via a fuzzer, the output of the maxsat solver has to
be gathered and analyzed to judge whether the solver had a flaw. This
script brings the basic functionality to execute a solver on an input file
and to collect the output, extract the relevant data points and do some
basic logic on a per-solver basis (for now).

With the lower defaults, the fuzzer can be called without parameters and
the generated formulas are still in a reasonable range and solve time.

Take care of uninitialized values. Drop solver output, as it can be
reproduced.

Instead of checking a single file, fuzz as many new files as specified
on the CLI.

To make the output useful, move most comments away from stderr for now.

Do not produce clauses with weight 0, as multiple solvers seem to not
accept it.

When a timeout is caught, a solver might determine a formula as
satisfiable. Hence, this combination is accepted and should be respected.

The third value t in the "p (wcnf) v c t line" is the sum of all weights
of all soft clauses plus some offset. All hard clauses in the formula have
to have a weight greater equal this value. Hence, pre-generate all weights
and calculate this sum, and then fix the weights of the hard clauses. This
way, valid MaxSAT input will be generated.

While being at it, also fix the way the iterations are printed.

To regenerate formulas based on the seed, the seed should be the only
parameter that is specified. Hence, make it the first parameter. While
being at it, also fix the value that are printed in the log.

Before generating the formula, print stats on the soft and hard clauses,
as well as the sum of the weights of the two classes.

When parsing the top value for plain CNF, the value was sometimes slightly
off. Change this behavior by initializing properly.

Make sure we are able to detect when the sum of weights for soft classes
exceeds the value of the top head entry.

When specifying a negative iteration number, the script will only call the
fuzzer, but it will not call the solvers on the generated files. This is
a simple way to generate a set of formulas with the fuzzer backend.

top has been initialized to -1, which is not required. Hence, move to 0.
Furthermore, all parameters are optional. Reflect this in the usage.

While any weight larger than the top value should result in a hard
clause, parsers of solver might actually compare the string, and hence
miss this case. Therefore, use the top value.

In case invalid input is generated, this rule might be violated again.