memory: close leaks in OLL
New encoders are stored and maintained in a vector via pointers.
However, after search is finished, these encoders are not freed,
but the vector is silently released. As the used vector is a data
type of MiniSat, the objects are not freed.

This problem is solved by moving the collection to a class member.
This way, releasing the memory can be done after search finished,
and can cover all returns of the search functions.