maxsat: add setSATbudget
Allow to specify the amount of SAT steps to be executed inside a MaxSAT
search. Those steps would be translated to conflicts in SAT.

In case the SAT solver consumes the budget, it will not return with
l_True or l_False, but l_Undef. This case should be handled in all
algorithms specially, so that the best can be made out of the intermediate
state.